Open api

+++ cloud-monitor-center/service/contact_group.go
@@ -15,6 +15,7 @@ import (
 	"code.cestc.cn/ccos-ops/cloud-monitor/common/util/snowflake"
 	"code.cestc.cn/ccos-ops/cloud-monitor/common/util/strutil"
 	"gorm.io/gorm"
+	"regexp"
 	"strconv"
 )
 
@@ -33,7 +34,7 @@ func NewContactGroupService(contactGroupRelService *ContactGroupRelService) *Con
 }
 
 func (s *ContactGroupService) PersistenceLocal(db *gorm.DB, param interface{}) (string, error) {
-	p := param.(form.ContactParam)
+	p := param.(*form.ContactParam)
 	//每个联系人最多加入5个联系组
 	if len(p.GroupBizIdList) >= constant.MaxContactGroup {
 		return "", errors.NewBusinessError("每个联系人最多加入" + strconv.Itoa(constant.MaxContactGroup) + "个联系组")
@@ -44,22 +45,21 @@ func (s *ContactGroupService) PersistenceLocal(db *gorm.DB, param interface{}) (
 		if strutil.IsBlank(p.GroupName) {
 			return "", errors.NewBusinessError("联系组名字不能为空")
 		}
+		if !s.checkGroupName(p.GroupName) {
+			return "", errors.NewBusinessError("联系组名字格式错误")
+		}
 		if len(p.ContactBizIdList) == 0 {
 			return "", errors.NewBusinessError("请至少选择一位联系人")
 		}
 		//联系组限制创建10个
-		var groupCount int64
-		global.DB.Model(&model.ContactGroup{}).Where("tenant_id = ?", p.TenantId).Count(&groupCount)
-		if groupCount >= constant.MaxGroupNum {
+		if s.dao.GetGroupCount(p.TenantId) >= constant.MaxGroupNum {
 			return "", errors.NewBusinessError("联系组限制创建" + strconv.Itoa(constant.MaxGroupNum) + "个")
 		}
 		//联系组名不可重复
-		var count int64
-		global.DB.Model(&model.ContactGroup{}).Where("tenant_id = ? AND name = ?", p.TenantId, p.GroupName).Count(&count)
-		if count >= 1 {
+		if s.dao.CheckGroupName(p.TenantId, p.GroupName, "") {
 			return "", errors.NewBusinessError("联系组名重复")
 		}
-		contactGroup, err := s.insertContactGroup(db, p)
+		contactGroup, err := s.insertContactGroup(db, *p)
 		if err != nil {
 			return "", err
 		}
@@ -77,13 +77,14 @@ func (s *ContactGroupService) PersistenceLocal(db *gorm.DB, param interface{}) (
 		if strutil.IsBlank(p.GroupName) {
 			return "", errors.NewBusinessError("联系组名字不能为空")
 		}
+		if !s.checkGroupName(p.GroupName) {
+			return "", errors.NewBusinessError("联系组名字格式错误")
+		}
 		//联系组名不可重复
-		var count int64
-		global.DB.Model(&model.ContactGroup{}).Where("tenant_id = ? AND biz_id != ? AND name = ?", p.TenantId, p.GroupBizId, p.GroupName).Count(&count)
-		if count >= 1 {
+		if s.dao.CheckGroupName(p.TenantId, p.GroupName, p.GroupBizId) {
 			return "", errors.NewBusinessError("联系组名重复")
 		}
-		contactGroup, err := s.updateContactGroup(db, p)
+		contactGroup, err := s.updateContactGroup(db, *p)
 		if err != nil {
 			return "", err
 		}
@@ -96,7 +97,7 @@ func (s *ContactGroupService) PersistenceLocal(db *gorm.DB, param interface{}) (
 			Data:     contactGroup,
 		}), nil
 	case enum.DeleteContactGroup:
-		contactGroup, err := s.deleteContactGroup(db, p)
+		contactGroup, err := s.deleteContactGroup(db, *p)
 		if err != nil {
 			return "", err
 		}
@@ -137,8 +138,10 @@ func (s *ContactGroupService) insertContactGroup(db *gorm.DB, p form.ContactPara
 }
 
 func (s *ContactGroupService) updateContactGroup(db *gorm.DB, p form.ContactParam) (*model.ContactGroup, error) {
-	var oldContactGroup model.ContactGroup
-	global.DB.Where("tenant_id = ? AND biz_id = ?", p.TenantId, p.GroupBizId).First(&oldContactGroup)
+	if strutil.IsBlank(p.GroupBizId) {
+		return nil, errors.NewBusinessError("联系组ID不能为空")
+	}
+	var oldContactGroup = s.dao.GetGroup(p.TenantId, p.GroupBizId)
 	if oldContactGroup == (model.ContactGroup{}) {
 		return nil, errors.NewBusinessError("该租户无此联系组")
 	}
@@ -158,10 +161,21 @@ func (s *ContactGroupService) updateContactGroup(db *gorm.DB, p form.ContactPara
 }
 
 func (s *ContactGroupService) deleteContactGroup(db *gorm.DB, p form.ContactParam) (*model.ContactGroup, error) {
+	//检验联系组是否存在
+	if !s.dao.CheckGroupId(p.TenantId, p.GroupBizId) {
+		return nil, errors.NewBusinessError("该租户无此联系组")
+	}
 	var contactGroup = &model.ContactGroup{
 		BizId:    p.GroupBizId,
 		TenantId: p.TenantId,
 	}
 	s.dao.Delete(db, contactGroup)
 	return contactGroup, nil
+}
+
+func (s *ContactGroupService) checkGroupName(groupName string) bool {
+	if regexp.MustCompile("^[a-zA-Z0-9_\u4e00-\u9fa5]{1,40}$").MatchString(groupName) {
+		return true
+	}
+	return false
 }
